Add a tutorial for using Ground Truth for object detection.
This commit adds a Jupyter Notebook with a tutorial for performing an
object detection labeling job using Amazon SageMaker Ground Truth. In
addition, it adds a file with support code for the notebook. This
support code consists of classes to organize the images and bounding
boxes produced during annotation and a function to calculate the image
mIoUs. Finally, the README is updated to include an entry for the new
tutorial.