Update setup.sh file to skip if the user doesn't have root access
* This change checks if current user has sudo access or not. If the
current user has sudo access then script is run as before and if
the current user doesn't have sudo access then skip the script.

* Run setup.sh file from Notebook with the change with and without root access
* With root access
```
The user has root access.
SageMaker instance route table setup is ok. We are good to go.
SageMaker instance routing for Docker is ok. We are good to go!
```

* Without root access
```
sudo: a password is required
The user does not have root access. Everything required to run the notebook is already installed and setup. We are good to go!
```