Fix copy of image data to work without signature
As per the documentation on CVDF, downloads should happen with the no-sign-request option. As published the notebook will fail to copy the images with a 403 error from the source about HeadObject calls being forbidden. Suppressing signatures in the client using config=Config(signature_version=UNSIGNED) will fail at the destination end with an authorization failure.

It appears that the S3 COPY operation requires the authentication on both ends to be the same, even with providing a separate s3 client for the Head/Get operation. So the only plausible workaround appears to be to do a get/put copy as proposed by this patch.