Updated hpo_xgboost_direct_marketing_sagemaker_python_sdk.ipynb for SageMaker SDK v2
Notebook needed updates to be compatible with the v2 of the SageMaker SDK, which is the new Default SDK version in SageMaker.

Issue #, if available:

Description of changes:
Notebook needed updates to be compatible with the v2 of the SageMaker SDK, which is the new Default SDK version in SageMaker.

Fixed:
- Use TrainingInput instead of s3_input.
- Use sagemaker.image_uris.retrieve instead of get_image_uri.
- Use instance_count and instance_type instead of train_instance_count and train_instance_type.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.