New SageMaker Clarify notebook to showcase supports for JSONLines format
SageMaker Clarify supports CSV format and JSONLines format; its existing
example notebook uses CSV format. This commit adds a new notebook which
is essentially the same as the old one but use JSONLines format instead.
A few updates are applied to the old one as well for cross-reference and
to fix minor issues. Both notebooks are added to processing index.rst .