Updated tensorflow_script_mode_training_and_serving.ipynb for SageMaker SDK v2
Notebook needed updates to be compatible with the v2 of the SageMaker SDK, which is the new Default SDK version in SageMaker.

Issue #, if available:

Description of changes:
Notebook needed updates to be compatible with the v2 of the SageMaker SDK, which is the new Default SDK version in SageMaker.

Fixed:

Use instance_count, instance_type, and distribution instead of train_instance_count, train_instance_type, and distributions.
Fixed sagemaker.Session().delete_endpoint(predictor.endpoint) to predictor.delete_endpoint().

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.