Amazon SageMaker Asynchronous Inference with Hugging face Model (#2910)
* Amazon SageMaker Asynchronous Inference with Hugging face Model A new near real-time Inference option for generating machine learning model predictions

This notebook provides an introduction on how to use the SageMaker Asynchronous inference capability with Hugging face models. This notebook will cover the steps required to create an asynchronous inference endpoint and test it with some sample requests.

* Updated with comments

* Amazon SageMaker Asynchronous Inference with Hugging face Model

This notebook provides an introduction on how to use the SageMaker Asynchronous inference capability with Hugging face models. This notebook will cover the steps required to create an Asynchronous inference endpoint and test it with some sample requests.

* Add files via upload

Commiting the changes done as per the feedbacks received by the reviewer.

* Add files via upload

updating aws doc link reference for all regions

* Updated the comment around upgrading sagemaker