Added new usecase for financial payment classification using Amazon SageMaker
Enriching financial transactions with the category of the transaction. This can be used as an intermediate set in fraud detection, personalization or anomaly detection. As well as a method to provide end users (e.g. customers of a bank) with an overview of their spending habits.

The notebook demonstrates how Amazon SageMaker can be used to train and deploy a XGBoost model, as well as the required underlying infrastructure. For this notebook a generated dataset is used where a payment consists of an amount, sender, receiver and time stamp.