Update fairness_and_explainability.ipynb
From working out the values of the post-training metrics from the model predictions, the values do not match the bias report values using the given logic that if the sum of local_explanations_out values (shap values) for given instance is > 0, then the prediction is True. However, they do match if the value of 0 is changed to the probability threshold of 0.8 used in the model.