fix: Amend PR 3760 with notebook outputs
The notebooks added by PR 3760 can take more than one hour to run
(mainly because the minimum interval of a monitoring schedule is
one hour). This commit keeps the outputs so that the customers
don't have to spend too much time if they just want to see what
the product does. It also makes the notebook pages on readthedocs
website a more complete document.