Append the default_bucket_prefix to s3 paths if one exists to hyperparameter_tuning, introduction_to_amazon_algorithms and sagemaker-pipelines sample notebooks
**Description**
This change checks the sagemaker session if there is a default_bucket_prefix set. If a default bucket prefix is specified, it is append it to the s3 path. This was done for hyperparameter_tuning, introduction_to_amazon_algorithms and sagemaker-pipelines sample notebooks. This ensures sample notebooks that utilize s3, works in SageMaker Unified Studio.