Update default_bucket and default_bucket_prefix to derive from sagemaker rather than sagemaker_core AND Append the default_bucket_prefix to s3 paths if one exists to sample notebooks (#4812)
* Append the default_bucket_prefix to s3 paths if one exists to sample notebooks

**Description**

This change checks the sagemaker session if there is a default_bucket_prefix set. If a default bucket prefix is specified, it is appended it to the s3 path. This ensures sample notebooks that utilize s3, works in SageMaker Unified Studio.

* Update default_bucket and default_bucket_prefix to derive from sagemaker rather than sagemaker_core

**Description**

In order to be compatible with SageMaker Unified Studio, the default_bucket and default_bucket_prefix must be derived from the sagemaker python sdk rather than the sagemaker_core python SDK.

---------

Co-authored-by: Marco Friaz <marfriaz@amazon.com>